feat: implement new CodeInput (from a library) and rename it from CodeInputLib to CodeInput + delete old CodeInput component
